+++ pkg/ui/workspaces/cluster-ui/src/index.ts
@@ -45,3 +45,4 @@ export * from "./text";
 export { util, api };
 export * from "./sessions";
 export * from "./timeScaleDropdown";
+export * from "./graphs";
